// frontend/src/types/types.ts
export interface Commercial {
  id: string;
  nom: string;
  prenom: string;
  email: string;
  telephone: string | null;
  equipeId?: string;
  managerId?: string;
  isAssigned?: boolean;
  currentMonthlyGoal?: number; // Ajouté pour correspondre au schéma Prisma
  historiques: { nbContratsSignes: number }[]; // Ajouté pour correspondre à l'API
  manager?: string;
  equipe?: string;
  classement?: number;
}
// ...
export type EnrichedCommercial = Commercial & {
  manager: string;
  equipe: string;
  classement: number;
  totalContratsSignes: number;
};